Consolidate cmdrunner tests and achieve 100% coverage
- Merged tests/test_cmdrunner_extra.py into tests/test_cmdrunner.py.
- Added test cases for ANSI color initialization (NO_COLOR, isatty).
- Added test cases for MinimalFormatter edge cases (uncolored levels, empty levelname).
- Improved configuration validation tests to cover all error branches.
- Verified 100% statement and branch coverage for cmdrunner.py.
